Add hack to make App model play nice with preview
When the UX compiler does its thing, it assumes the `App` tag is an instance of `AppBase`.
However, when running in preview, the app tag is actually an instance of FakeApp (a Panel).
This exploits the bytecode VM's dynamically typed nature, relying on FakeApp quacking like an App.

Now, `Model` has two attached property setters, one for AppBase, and one for Visual.
These previously had the same name, which means:

- The UX compiler would generate code that calls `SetModel(AppBase, IExpression)`
- When the bytecode is run, it will attempt to call `SetModel(fakeApp, ...)`
- However, since FakeApp is a Visual, it will select the overload that takes a Visual.

But what we actually want to do is to set the app model
on the actual AppBase instance that FakeApp holds a reference to.

This simply renames the attached property setter for App to SetAppModel, making them unambiguous.
Now we can handle the case of `SetAppModel(FakeApp, ...)` inside the simulator app instead.