ScriptException: Truncate long source lines
It is quite common to minify JavaScript libraries in production,
resulting in the entire source code appearing on the same line.
This means that if an exception is thrown, the log will be filled with
the entirety of the minified file. This is not very helpful, so truncate
unreasonably long lines of source code.

The number 600 was chosen somewhat arbitrarily and is up for discussion.